Rework run and venv setup scripts

+++ set-up-venv.sh
@@ -1,32 +1,28 @@
 #!/usr/bin/env bash
 
-sudo apt-get update
+set -eou pipefail
 
-vnv=virtualenv
+cd "$(dirname "$0")"
 
-if ! command -v virtualenv > /dev/null; then
-  sudo apt-get install python3-distutils
-  wget -O /tmp/virtualenv.pyz https://bootstrap.pypa.io/virtualenv.pyz
-  vnv="python3 /tmp/virtualenv.pyz"
+if [ -x venv/bin/ansible-playbook ]; then
+  source venv/bin/activate
+else
+  vnv=virtualenv
+
+  if ! command -v virtualenv > /dev/null; then
+    sudo apt-get install -y python3-distutils
+    if [ ! -f /tmp/virtualenv.pyz ]; then
+      wget -O /tmp/virtualenv.pyz https://bootstrap.pypa.io/virtualenv.pyz
+    fi
+    vnv="python3 /tmp/virtualenv.pyz"
+  fi
+
+  $vnv venv -p python3
+  source ./venv/bin/activate
+
+  sudo apt-get update
+  # needed by psutil
+  sudo apt-get install -y python3-dev
+
+  pip install -r requirements.txt
 fi
-
-$vnv venv -p python3
-source ./venv/bin/activate
-
-# needed by psutil
-sudo apt-get install python3-dev
-
-mkdir py-apt
-sudo chown _apt py-apt
-
-cd py-apt
-
-sudo apt-get download python3-apt
-dpkg -x python3-apt_*.deb python-apt
-cp -r ./python-apt/usr/lib/python3/dist-packages/* ../venv/lib/python3.8/site-packages/
-
-cd ..
-
-rm -rf py-apt
-
-pip install wheel psutil "ansible>=2.9<2.11"
